+++ examples/dummy.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+hermes-cli -o ./ --config_file dummy.yml --experiment project=Test type=Dummy trial=0