+++ src/hermes/base/nodes/pipeline.py
@@ -26,17 +26,17 @@
 # ############
 
 from abc import abstractmethod
+from collections import OrderedDict
 import threading
 import zmq
 
 from hermes.utils.node_utils import search_node_class
 from hermes.utils.msgpack_utils import deserialize, serialize
-from hermes.utils.dict_utils import *
-from hermes.utils.zmq_utils import *
+from hermes.utils.zmq_utils import CMD_END, CMD_EXIT, DNS_LOCALHOST, PORT_BACKEND, PORT_FRONTEND, PORT_SYNC_HOST, PORT_KILL
 
 from hermes.base.stream import Stream
-from hermes.base.storage import Storage
-from hermes.base.nodes import Node
+from hermes.base.storage.storage import Storage
+from hermes.base.nodes.node import Node
 from hermes.base.nodes.pipeline_interface import PipelineInterface
 from hermes.base.nodes.producer_interface import ProducerInterface